// src/app/models/event.ts
export class EventModel {
  id?: number;
  ministerio: string;
  titulo: string;
  modalidad: string;
  optionplace: string;
  place: string;
  tipofecha : string;
  fechasingle? : string;
  fechaini? : string;
  fechafin? : string;
  horaini : string;
  horafin? : string;
  descripcion : string;
  encargado?: string;
}